begin major cleanup and refactoring
start with the entry point - main.cpp + CADventory.cpp

main is a bare-bones entry point for the qtapplication and
execution. de-couple CADventory class from the main QApplication.
Eventually we could fully isolate CADventory from qt (since we've put it
in a QObject), but in the meantime this allows separation from GUI / CLI
running.

first pass is intended to move code (whitespace / function locations /
etc.) but fundamental logic should not be effected yet. Instead,
sprinkle in a few TODO notes for likely point of improvement.