start roughing in a JobService
JobService will eventually be owned by the parent CADVentory 'app' and
either be a manager - in charge of managing the fs+sqlite state. Or a
worker - in charge of actually processing files (this will eventually
succeed the indexWorker).

No actual logic yet, but make sure we can create an abstract job service
with typical characteristics run / stop / get updates.