work on sqlite based jobqueue
FileJobQueue doesn't seem to be cutting it in terms of reliability
(std::filesystem::rename is not cross platform atomic), and speed. Start
roughing in simple logic for create / claim / finish. Right now, testing
and setup is for a unique queue.db file, but we may want to integrate
with the existing model.db sqlite databse (maybe as a foreign key) in
the future.