be a little more gentle in polling
don't create jobs while we still have jobs to do since there's a high
likelyhood it's already in there and we're just wasting db/filesystem
checks.
limit our scan intervals for the job queue and model repo to not hammer
fs checks.
protect the whole loop body in lazy try/catch so we dont kill the thread
if the db is throwing 'busy' due to syncing other workers.