remove files without objects from selection
if we get an empty or corrupt file with no valid tops objects, existing
code will skip it, never mark it processed, and infinitely try to
process it on each JobWorker loop.

Instead, if we know the file has no valid tops objects, go ahead and
invalidate the selection status.