start converting gist report into jobQueue jobs
create a job structure so we can dispurse the gist generation work.

ReportGenerationWorker no longer does the actual report work - it instead
queues a gist_report job.
However, because of this we no longer have an easy way to track
progress.. We may want to add some mechanism for that in the future.

Also, our use of the JobQueue is a little 'hacky' - we write a json blob
into the sourcePath field and unpack it in the gist handler. We can /
should create an optional metadata field in the job which could be
useful for situations like this (but for now this works).