fix build for non-gui
if we've built qt without gui, we don't have access to qt image,
painter, or pdfwriter. For now, just #if that logic out of the Gist
Report handler, and rely on an instance running gui to handle the final
pdf report assembly.

Eventually, we should just ditch qt for the pdf assembly since it's not
relevant and instead include a small pdf creation library. But, since
there's not a way to even kick off the gist report generation without a
gui build right now, this is a decent duct-tape fix.