better update properties
previous behvaior would hit the model on all property changes (each
keystroke!). Instead, collect the changes in memory and only hit the
model db on 'ok'.

Intentionally freeze the window while the db is being hit so it's
painfully obvious to the user that something is happening and to not
move on until everything is synced. On most instances this is not
noticeable - but on congested / slow fs we don't want to collide db
state if we're updating models faster than the fs can write.