expose setter for SimpleFileLock timeout
It's a little hoaky to expose it like this, but SQLJobQueueTest is
intermittently timing out of create jobs. In 'real' use a timeout is no
big deal - the job is dumped and the claim will be tried again by
another worker/thread. But for the tests everything is one-shot so a
timeout means a failure.